exceptions utils
util class created to read user input
Book child classes are moved to individual classes to make public in order to create instance out of the package
Some toStrings methods are created

+++ Library/src/service/UserService.java
@@ -19,7 +19,7 @@ public class UserService {
     }
     
     // Find a user by ID
-    public User findUserById(int id) {
+    public User findUserById(long id) {
         // use of for loop
         for (User user : users) {
             if (user.getId() == id) {
